api/utils.py
import io

from PIL import Image


def load_image_from_bytes(image_bytes: bytes) -> Image.Image:
    try:
        image = Image.open(io.BytesIO(image_bytes)).convert("RGB")
    except Exception as exc:
        raise ValueError("No se pudo leer la imagen enviada") from exc
    return image
